refactor(staking-cli): move the L1 block range config into the l1 module
It is provider configuration, not command logic, and `stake-table` no longer
has to reach into the `stake-table-entry` module for it.
