+++ staking-cli/src/entry.rs
@@ -41,6 +41,7 @@ use hotshot_types::{addr::NetAddr, light_client::StateVerKey, signature_key::BLS
 use serde::Serialize;
 
 use crate::{
+    l1::{BLOCK_RANGE_VAR, DEFAULT_BLOCK_RANGE, configured_block_range},
     output::{Esp, output_success},
     parse::Commission,
 };
@@ -292,27 +293,6 @@ pub async fn fetch_stake_table_entry(
         approximate,
     })
 }
-
-/// Blocks per request when the range has to be split. Matches the default of the environment
-/// variable that overrides it, which is the range capped providers usually allow.
-const DEFAULT_BLOCK_RANGE: u64 = 10_000;
-
-/// The block range the user pinned for their provider, if any.
-///
-/// Set means "my provider caps the range at this", so the single request is skipped rather than
-/// spent on a rejection.
-pub(crate) fn configured_block_range() -> Option<u64> {
-    let value = std::env::var_os(BLOCK_RANGE_VAR)?;
-    match value.to_str().and_then(|value| value.parse().ok()) {
-        Some(range) if range > 0 => Some(range),
-        _ => {
-            tracing::warn!("ignoring {BLOCK_RANGE_VAR}: expected a positive integer");
-            None
-        },
-    }
-}
-
-pub(crate) const BLOCK_RANGE_VAR: &str = "ESPRESSO_L1_EVENTS_MAX_BLOCK_RANGE";
 
 /// Fetch a filter's logs in one request, splitting the range only if that is refused.
 ///
